recipes-bsp: pmu-firmware: fix zu1 PMU firmware
The power is brought up and shutdown on both the Ultra96-V2 and ZU1
board by the ‘EN_SEQ_PL’ signal from the on/off controller.  Here is how
that works:

1. To power on the board a short press of the power switch asserts the
‘EN_SEQ_PL’ signal, which in turn enables the power supplies to the ZU+
device and the rest of the board.
2. To turn off the board a short press of the power switch asserts the
‘MIO26_PWR_INT’ signal.  This signal is received by the ZU+ device and
the PMU firmware then initiates a shutdown of the OS and then assert the
‘MIO34_POWER_KILL_N’ signal which the on/off controller then recognizes
to deassert the EN_SEQ_PL signal and turn off the power supplies.
3. The board can also powered off by issuing the ‘poweroff’ command in
Linux to halt the OS and the PMU firmware asserts the
‘MIO34_POWER_KILL_N’ signal which the on/off controller then recognizes
to deassert the EN_SEQ_PL signal and turn off the power supplies.

Before this change, the PMU firmware wasn't asserting
‘MIO34_POWER_KILL_N’ at power off